refactor(chaos): extract the instance-stall detector and its view out of ProgressProbe
ProgressProbe had grown to twelve hundred lines, and the instance-progress detector was most of
the growth: the bound and its rationale, the fleet view and its factory, the marks, the busy and
dump bookkeeping, the tokens, the sampler, the busy-worker count and the thread dump - one
detector's whole state machine nested inside the class that also owns the sampler thread and
three other detectors. InstanceStallDetector now holds all of it, InstanceProgressView is its
own file, and the probe keeps thin delegates by the names everything cites -
sampleInstanceProgress, instanceProgressSnapshot, withInstanceProgress, withThreadDumpSource,
INSTANCE_STALL_BOUND, getPeakInstanceStallMs - so InstanceStallProbeIT, ChaosScenarioBase,
AmbientProbeExtension and every record that names those anchors still resolve.

Findings still land through the probe: the detector is constructed with the probe's violate and
observe as consumers, so the record rule - mode-dependent log level, the observation marker
bin/chaos-test.sh counts - stays in one place. A pure move: every javadoc travelled with its
member, cross-file links were repointed, and InstanceStallProbeIT passes unchanged in substance.
