+++ parallel-consumer-core/src/test-integration/java/bz/stub/parallelconsumer/integrationTests/chaostests/ChaosScenarioBase.java
@@ -440,7 +440,7 @@ abstract class ChaosScenarioBase extends BrokerIntegrationTest<String, String> i
      * LIVE fleet view - a supplier, because JOIN_NEW grows the fleet mid-run. */
     protected void startRun(ProgressProbe probe, ChaosConductor conductor) {
         probe.withInstanceProgress(() -> conductor.getFleet().stream()
-                .map(ProgressProbe.InstanceProgressView::of)
+                .map(InstanceProgressView::of)
                 .collect(Collectors.toList()));
         probe.start();
         conductor.start();
