+++ parallel-consumer-core/src/test-integration/java/bz/stub/parallelconsumer/integrationTests/chaostests/InstanceStallProbeIT.java
@@ -39,7 +39,7 @@ class InstanceStallProbeIT {
     private static final Instant T0 = Instant.parse("2026-01-01T00:00:00Z");
 
     /** Mutable scripted view - the test flips its fields between samples. */
-    private static final class FakeInstance implements ProgressProbe.InstanceProgressView {
+    private static final class FakeInstance implements InstanceProgressView {
         final int id;
         boolean live = true;
         long queued;
@@ -47,7 +47,7 @@ class InstanceStallProbeIT {
         long workResultsReturned;
         Object incarnation = new Object();
         /** Unknown by default, so every test written before the busy-worker rule still exercises the old one. */
-        int busyWorkers = ProgressProbe.BUSY_WORKERS_UNKNOWN;
+        int busyWorkers = InstanceStallDetector.BUSY_WORKERS_UNKNOWN;
 
         FakeInstance(int id) {
             this.id = id;
@@ -92,7 +92,7 @@ class InstanceStallProbeIT {
     /** A probe with only the instance-progress detector armed - the ctor's null kcu is legal because
      * the sampler thread is never started. */
     private static ProgressProbe probeWatching(FakeInstance... instances) {
-        List<ProgressProbe.InstanceProgressView> views = new ArrayList<>(Arrays.asList(instances));
+        List<InstanceProgressView> views = new ArrayList<>(Arrays.asList(instances));
         return ProgressProbe.forSeamTest("test-group", "test-topic")
                 .withInstanceProgress(() -> views);
     }
@@ -223,7 +223,7 @@ class InstanceStallProbeIT {
 
     /**
      * A firing takes ONE thread dump, and the default configuration is the case that needs saying so:
-     * {@link ProgressProbe#INSTANCE_STALL_DUMP_AFTER} defaults to
+     * {@link InstanceStallDetector#INSTANCE_STALL_DUMP_AFTER} defaults to
      * {@link ProgressProbe#INSTANCE_STALL_BOUND} itself, so the first sample past the bound satisfies
      * the early-dump condition and the violation condition on the same {@code stalledMs}. Taking the
      * dump in both branches paid a second {@code ThreadMXBean#getThreadInfo(ids, true, true)} - the
@@ -369,7 +369,7 @@ class InstanceStallProbeIT {
         Thread mine = parked("pc-pool-3-thread-2-PC-1", release);
         Thread lookalike = parked("pc-control-PC-14", release);
         try {
-            String dump = ProgressProbe.instanceThreadDump(1);
+            String dump = InstanceStallDetector.instanceThreadDump(1);
 
             assertWithMessage("the accused instance's own thread, with its state and a frame to read")
                     .that(dump).contains("\"pc-pool-3-thread-2-PC-1\" WAITING");
@@ -473,18 +473,18 @@ class InstanceStallProbeIT {
             await().atMost(Duration.ofSeconds(5)).until(() -> lookalike.getActiveCount() == 1);
 
             assertWithMessage("two tasks running, the third worker never created: two busy")
-                    .that(ProgressProbe.busyWorkersOf(4242)).isEqualTo(2);
+                    .that(InstanceStallDetector.busyWorkersOf(4242)).isEqualTo(2);
             assertWithMessage("-PC-4242 is a substring of -PC-42421; the lookalike's busy worker is its own")
-                    .that(ProgressProbe.busyWorkersOf(42421)).isEqualTo(1);
+                    .that(InstanceStallDetector.busyWorkersOf(42421)).isEqualTo(1);
             assertWithMessage("an instance with no pool threads at all counts nothing busy")
-                    .that(ProgressProbe.busyWorkersOf(424_242)).isEqualTo(0);
+                    .that(InstanceStallDetector.busyWorkersOf(424_242)).isEqualTo(0);
 
             release.countDown();
             // awaited, not asserted: a worker's active flag clears a few instructions before it is
             // back inside getTask, and the count reads the frame, not the flag
             await().alias("released: both workers parked between tasks, none busy")
                     .atMost(Duration.ofSeconds(5))
-                    .until(() -> ProgressProbe.busyWorkersOf(4242) == 0);
+                    .until(() -> InstanceStallDetector.busyWorkersOf(4242) == 0);
         } finally {
             release.countDown();
             pool.shutdownNow();
@@ -509,7 +509,7 @@ class InstanceStallProbeIT {
      */
     @Test
     void threadDumpSaysSoWhenTheInstanceHasNoThreads() {
-        assertThat(ProgressProbe.instanceThreadDump(999_999)).contains("no threads named *-PC-999999");
+        assertThat(InstanceStallDetector.instanceThreadDump(999_999)).contains("no threads named *-PC-999999");
     }
 
     private static Thread parked(String name, CountDownLatch until) {
